Day 33/iss_tracker/pyproject.toml
[tool.poetry]
name = "apis"
version = "0.1.0"
description = ""
authors = ["Eddy Cedeño <miguepoke28@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
